requirements.txt
# Data Sources
tweepy
praw
yfinance
requests

# Sentiment & NLP
vaderSentiment
transformers
torch
spacy
textblob

# VectorDB (optional, only if you're storing embeddings)
chromadb

#  LLM & APIs
groq
openai  # (Optional fallback, not needed if only using Groq)

# Backend & UI
flask
streamlit
python-dotenv

# Data Handling
pandas
matplotlib
scikit-learn
tqdm
